Split the struct_opts definition out
Now files that reference it can include opts.h
